feat(data): format sqlite + minor
